SetVariants was extracted into separated file [#METR-23882]
